SERVER-13874 Fix MSVC signed/unsigned mismatch warning

+++ src/mongo/bson/timestamp.h
@@ -59,7 +59,7 @@ namespace mongo {
         Timestamp(Seconds s, unsigned increment) : Timestamp(s.count(), increment) {}
 
         Timestamp(unsigned a, unsigned b) : i(b), secs(a) {
-            dassert(secs <= std::numeric_limits<int>::max());
+            dassert(secs <= static_cast<unsigned>(std::numeric_limits<int>::max()));
         }
 
         Timestamp() = default;
